requirements.txt
asyncpg>=0.20.1,<1.0.0
discord.py>=1.3.3,<2.0.0
toml>=0.10.0,<1.0.0
